requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pymongo==4.5.0
python-dotenv==1.0.0
certifi==2023.7.22
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
groq>=1.2.0
